sales/pyproject.toml
[project]
name = "sales"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "faker>=37.4.0",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "scipy>=1.16.0",
]
